+++ usage/completion/bash/bash_test.go
@@ -2,7 +2,7 @@ package bash
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
-//                         Copyright (c) 2020 ESSENTIAL KAOS                          //
+//                         Copyright (c) 2021 ESSENTIAL KAOS                          //
 //      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -17,11 +17,11 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-const _RESULT = `# Completion for test
+const _RESULT_FILES = `# Completion for test
 # This completion is automatically generated
 
 _test() {
-  local cur prev cmds opts
+  local cur prev cmds opts show_files
 
   COMPREPLY=()
   cur="${COMP_WORDS[COMP_CWORD]}"
@@ -29,6 +29,8 @@ _test() {
 
   cmds="print clean"
   opts="--option-aaa --option-bbb --option-ccc"
+  show_files="true"
+  file_glob="[sr]pm"
 
   case $prev in
     clean)
@@ -44,28 +46,105 @@ _test() {
     return 0
   fi
 
-  COMPREPLY=($(compgen -W '$(_test_filter "$cmds" "$opts")' -- "$cur"))
+  _test_filter "$cmds" "$opts" "$show_files" "$file_glob"
 }
 
 _test_filter() {
-  if [[ -z "$1" ]] ; then
-    echo "$2" && return 0
-  fi
+  local cmds="$1"
+  local opts="$2"
+  local show_files="$3"
+  local file_glob="$4"
 
   local cmd1 cmd2
 
-  for cmd1 in $1 ; do
+  for cmd1 in $cmds ; do
     for cmd2 in ${COMP_WORDS[*]} ; do
       if [[ "$cmd1" == "$cmd2" ]] ; then
-        echo "$2" && return 0
+        if [[ -z "$show_files" ]] ; then
+          COMPREPLY=($(compgen -W "$opts" -- "$cur"))
+        else
+          _filedir "$file_glob"
+        fi
+
+        return 0
       fi
     done
   done
 
-  echo "$1" && return 0
+  if [[ -z "$show_files" ]] ; then
+    COMPREPLY=($(compgen -W "$cmds" -- "$cur"))
+    return 0
+  fi
+
+  _filedir "$file_glob"
 }
 
-complete -F _test test -o nosort
+complete -F _test test -o filenames
+`
+
+const _RESULT_NO_FILES = `# Completion for test
+# This completion is automatically generated
+
+_test() {
+  local cur prev cmds opts show_files
+
+  COMPREPLY=()
+  cur="${COMP_WORDS[COMP_CWORD]}"
+  prev="${COMP_WORDS[COMP_CWORD-1]}"
+
+  cmds="print clean"
+  opts="--option-aaa --option-bbb --option-ccc"
+  show_files=""
+  file_glob=""
+
+  case $prev in
+    clean)
+      opts="--option-ddd --option-eee"
+      COMPREPLY=($(compgen -W "$opts" -- "$cur"))
+      return 0
+      ;;
+
+  esac
+
+  if [[ $cur == -* ]] ; then
+    COMPREPLY=($(compgen -W "$opts" -- "$cur"))
+    return 0
+  fi
+
+  _test_filter "$cmds" "$opts" "$show_files" "$file_glob"
+}
+
+_test_filter() {
+  local cmds="$1"
+  local opts="$2"
+  local show_files="$3"
+  local file_glob="$4"
+
+  local cmd1 cmd2
+
+  for cmd1 in $cmds ; do
+    for cmd2 in ${COMP_WORDS[*]} ; do
+      if [[ "$cmd1" == "$cmd2" ]] ; then
+        if [[ -z "$show_files" ]] ; then
+          COMPREPLY=($(compgen -W "$opts" -- "$cur"))
+        else
+          _filedir "$file_glob"
+        fi
+
+        return 0
+      fi
+    done
+  done
+
+  if [[ -z "$show_files" ]] ; then
+    COMPREPLY=($(compgen -W "$cmds" -- "$cur"))
+    return 0
+  fi
+
+  _filedir "$file_glob"
+}
+
+complete -F _test test 
 `
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -81,8 +160,11 @@ var _ = Suite(&BashSuite{})
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func (s *BashSuite) TestGenerator(c *C) {
-	data := Generate(genTestUsageInfo(), "test")
-	c.Assert(data, Equals, _RESULT)
+	completion := Generate(genTestUsageInfo(true), "test", "[sr]pm")
+	c.Assert(completion, Equals, _RESULT_FILES)
+
+	completion = Generate(genTestUsageInfo(false), "test")
+	c.Assert(completion, Equals, _RESULT_NO_FILES)
 }
 
 func (s *BashSuite) TestAuxi(c *C) {
@@ -94,8 +176,14 @@ func (s *BashSuite) TestAuxi(c *C) {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func genTestUsageInfo() *usage.Info {
-	info := usage.NewInfo("")
+func genTestUsageInfo(withFiles bool) *usage.Info {
+	var info *usage.Info
+
+	if withFiles {
+		info = usage.NewInfo("", "spec-file")
+	} else {
+		info = usage.NewInfo("")
+	}
 
 	info.AddCommand("print", "Print command")
 	info.AddCommand("clean", "Clean command")