+++ usage/completion/bash/examples_test.go
@@ -2,7 +2,7 @@ package bash
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
-//                         Copyright (c) 2020 ESSENTIAL KAOS                          //
+//                         Copyright (c) 2021 ESSENTIAL KAOS                          //
 //      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -30,5 +30,5 @@ func ExampleGenerate() {
 	info.AddOption("h:help", "Show this help message")
 	info.AddOption("v:version", "Show version")
 
-	fmt.Println(Generate(info, "app"))
+	fmt.Println(Generate(info, "app", "txt"))
 }