+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 12.13.0
+
+* `[usage/completion/bash]` Improved completion generation
+* `[usage/completion/zsh]` Improved completion generation
+* `[usage/completion/bash]` Fixed bug with showing files with autocomplete
+* `[usage/completion/zsh]` Fixed bug with showing files with autocomplete
+
 ### 12.12.0
 
 * `[timeutil]` Added method `PrettyDurationInDays` for rendering pretty duration in days